LIBDRUM-657. WIP for CAS authentication
See https://confluence.umd.edu/display/~dsteelma/Jira+-+LIBDRUM-657
for discussion.

This WIP basically only shows that successful CAS authentication is
possible.

Also, the "minimal" set of changes needed to make CAS authentication
work has not been determined, and are probably not anywhere near the
changes needed for a "final" implementation. For example, ideally the
changes in the "CASLoginFilter" and "WebSecurityConfiguration" files
should be moved to the "additions" modules, and not be in DSpace core.

https://issues.umd.edu/browse/LIBDRUM-657