LIBDRUM-696. Commented out UMD item submission customizations
The UMD-specific changes to the item submission workflow have not yet
been implemented in DSpace 7, so commenting out the customizations in
the "dspace/config/item-submission.xml" file, so at least the
default item submission workflow can work properly.

Work on the UMD customized workflows are in Jira as LIBDRUM-681
(for the Library Award collection) and LIBDRUM-684  (for the MHHEA
collection) will presumable reverse these changes at some point.

https://issues.umd.edu/browse/LIBDRUM-696