LIBDRUM-669. Refactor Ldap interface - Step 2
Refactored "get<Attribute>" methods into an LdapInfo object that is
stored in the request session, instead of the full LdapImpl object.

Modified "checkUid" method of "Ldap" interface to return an
"LdapInfo" object, instead of just a boolean. A null object means
the given user id was not found in LDAP.

https://issues.umd.edu/browse/LIBDRUM-669