LIBDRUM-669. Refactored LdapInfo "getUnits" to "getOrganizationalUnits"
The "getUnits" method was returning a list of LDAP "ou" entries for
the user, so renamed the method to "getOrganizationalUnits" to make this
clearer. Also made "protected" instead of "public" since it is not
used outside of the class.

Renaming this method will allow a new "getUnits" method to return a
list of the DSpace Units that the user is a member of to be created.

https://issues.umd.edu/browse/LIBDRUM-669