LIBDRUM-678. Endpoint to retrieve EmbargoDTO for bitstream UUID
Extended the EmbargoRestController/EmbargoDTOService/EmbargoDTODAO
classed to return the EmbargoDTO for a single bitstream, specified
by the bitstream's UUID.

If the UUID provided is not a bitstream, an empty map is returned.

The UUID endpoint should allow anonymous access, to enable Angular to
query for the embargo end date (used for "Restricted Access"
functionality).

https://umd-dit.atlassian.net/browse/LIBDRUM-678