LIBDRUM-678. Modified Bitstream to include restricted access flag
Added "isEmbargoed(DSpaceObject object)" method to "EmbargoDTOService"
and "EmbargoDTOServiceImpl" to indicate whether or not a particular
DSpaceObject (i.e., a bitstream) is embargoed. Placed in the
"EmbargoDTOService", so all the "ETD Embargo" retrieval code is in
the same place.

Added a "restrictedAccess" flag to the "BitstreamRest" model passed
between the back-end and front-end to indicate whether the bitstream
is embargoed.

Modified the "BitstreamConverter" class that generates the
"BitstreamRest" model to populate the "restrictedAccess" flag.

https://umd-dit.atlassian.net/browse/LIBDRUM-678