LIBDRUM-769. Refactored LdapServiceImpl to use in-memory cache
Added a “LdapServerDelegate” class, to actually perform the query
against the LDAP server. Primarily intended to enable tests to replace
the actual LDAP server calls with a mock implementation.

Added an in-memory persistent cache using the
"org.apache.commons.collections4.map.PassiveExpiringMap” class, in a
static “ldapQueryCache” variable, which is shared across all
LdapServiceImpl instances.

Both the LdapServer “delegate” and “ldapQueryCache” variables were given
“protected” access to enable unit tests.

Modified the “queryLdap” method to first check the “ldapQueryCache” for
the entry, and only querying the actual LDAP server if the entry is not
found.

Added unit tests to test cache handling.

https://umd-dit.atlassian.net/browse/LIBDRUM-769