LIBDRUM-769. Refactored CASAuthentication special groups handling
Added a “getLdap” helper method, the returns a
“edu.umd.lib.dspace.authenticate.impl.Ldap” object, based on the contents
of the CAS_LDAP (“cas.ldap”) attribute in the request’s session object,
or if the user is being impersonated. In case of imperonations, the
“queryLdap” method will be called to generate the “Ldap” object.

Renamed a “queryLdapService” method to simpler "queryLdap".

Updated the unit tests to impersonate based on
"Context.isContextUserSwitched()" method, instead of "X-On-Behalf-Of"
header in the request.

https://umd-dit.atlassian.net/browse/LIBDRUM-769