LIBDEVOPS-1754. POC for using Minio for S3 Storage in local development
This proof-of-concept code modifies the "S3BitStoreService" class with
a hard-coded "http://host.docker.internal:9000" URL to enable it to be
used with Minio object storage container running in Docker on the local
workstation (see https://min.io/docs/minio/container/index.html). TerpAI
assisted in generating these code changes.

In "bitstore.xml", removed the commented out code for assetstore1 (which
is not used), and added an additional assetstore5, which uses the
AWS S3 functionality.

In "local.cfg.EXAMPLE", added the configuration settings needed to
use the AWS S3 functionality.

https://umd-dit.atlassian.net/browse/LIBDEVOPS-1754