LIBDRUM-959. WIP - Change "drum-0" logs to STDOUT and JSON
THESE CHANGES TO NOT CURRENTLY APPEAR TO BE WORKING

Changed the UmdTomcatWebServerFactoryCustomizer" in an attempt to
add a "logfile: access.log" JSON attribute to the existing Tomcat log,
so that it can be queryable in Splunk.

Changed the "load-etd" script to use an XML Log4J configuration.
NOTE: The ETD Loader also stores and mails the log, so further
consideration of this functionality will be needed.

These changes are in conjunction with changes in the "k8s-drum"
repository.

https://umd-dit.atlassian.net/browse/LIBDRUM-959